requirements.txt
psycopg2-binary
python-dotenv
bs4
requests
selenium
# transformers
# pandas
# wordcloud
# matplotlib